Fix #13: use the correct database on all threads.
