+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2017-05-30 14:32:18 -0400 using RuboCop version 0.47.1.
+# on 2017-06-03 12:33:20 -0400 using RuboCop version 0.47.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -26,7 +26,7 @@ Lint/UselessAccessModifier:
 
 # Offense count: 3
 Metrics/AbcSize:
-  Max: 33
+  Max: 37
 
 # Offense count: 5
 # Configuration parameters: CountComments, ExcludedMethods.
@@ -37,7 +37,7 @@ Metrics/BlockLength:
 Metrics/CyclomaticComplexity:
   Max: 8
 
-# Offense count: 44
+# Offense count: 46
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
