+++ spec/spec_helper.rb
@@ -5,9 +5,7 @@ require 'rspec'
 require 'mongoid'
 require 'timecop'
 
-Mongoid.configure do |config|
-  config.connect_to('mongoid-collection-snapshot_test')
-end
+Mongoid.load!("#{File.dirname(__FILE__)}/config/mongoid#{ENV['MONGOID_VERSION'] || 6}.yml", :test)
 
 require File.expand_path('../../lib/mongoid-collection-snapshot', __FILE__)
 Dir["#{File.dirname(__FILE__)}/models/**/*.rb"].each { |f| require f }
