--- t/dump/widgets-and-gadgets/widgets_and_gadgets.snapshot.metadata.json
@@ -1,1 +1,0 @@
-{"options":{},"indexes":[{"v":1,"key":{"_id":1},"name":"_id_","ns":"widgets-and-gadgets.widgets_and_gadgets.snapshot"}]}
